refactor - consolidate auth types and relocate misplaced components
- Merge currentSessionTypes into authTypes — one type file per feature
- Delete currentSessionTypes.ts; update all imports accordingly
- Move AuthConfirm and AuthIntro from components/ui/ to features/auth/components/
- Move PasswordStrengthHint from components/ui/ to components/common/ (used by auth and profile)
- Update all import paths across signin, signup, forgot password, reset password, and profile
