+++ apps/web/src/components/common/PasswordStrengthHint.tsx
