+++ apps/web/src/features/auth/authTypes.ts
@@ -73,3 +73,22 @@ export type CurrentSession = {
 export type CurrentSessionResponse = {
   data: CurrentSession;
 };
+
+export type CurrentSessionStatus =
+  | "loading"
+  | "guest"
+  | "authenticated"
+  | "error";
+
+export type CurrentSessionContextValue = {
+  status: CurrentSessionStatus;
+  currentSession: CurrentSession | null;
+  user: SessionUser | null;
+  permissions: UserPermissions | null;
+  isLoading: boolean;
+  isAuthenticated: boolean;
+  isAdmin: boolean;
+  error: string | null;
+  refreshSession: () => Promise<CurrentSession | null>;
+  signOut: () => Promise<void>;
+};
