+++ apps/web/src/features/auth/session/index.ts
@@ -3,4 +3,4 @@ export { useCurrentSession } from "@/features/auth/session/useCurrentSession";
 export type {
   CurrentSessionContextValue,
   CurrentSessionStatus,
-} from "@/features/auth/session/currentSessionTypes";
+} from "@/features/auth/authTypes";
